api/pyproject.toml
[project]
name = "neurogram"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Michael Chernyakoff", email = "chernyakoff@gmail.com"},
]
dependencies = [
    "tortoise-orm[accel,asyncpg]>=0.25.1",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "aiopath>=0.7.7",
    "cyclopts>=3.19.0",
    "watchfiles>=1.1.0",
    "python-dateutil>=2.9.0.post0",
    "aioclock>=0.3.0",
    "uvicorn[standard]>=0.35.0",
    "fastapi>=0.116.1",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "emojis>=0.7.0",
    "hatchet-sdk>=1.18.1",
    "python-multipart>=0.0.20",
    "aerich>=0.9.1",
    "autoflake>=2.3.1",
    "aioboto3>=15.1.0",
    "tortoise-serializer>=1.6.0",
    "html-to-markdown>=2.9.1",
    "openrouter>=0.1.1",
    "httpx[socks]>=0.28.1",
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false


[tool.taskipy.tasks]
app = "python -m app"
dev = "python -m app.dev"
migration = "aerich migrate"
migrate = "aerich upgrade"
unmigrate = "aerich downgrade"
autoflake = "autoflake ./app"
cli = "python -m app.cli"


[tool.aerich]
tortoise_orm = "app.config.tortoise_config"
location = "./migrations"
src_folder = "."

[tool.autoflake]
ignore-init-module-imports = true
remove-all-unused-imports = true
recursive = true
in-place = true

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
]
